MajorFlask/requirements.txt
gradio>=3.38
groq
python-dotenv
pydub
SpeechRecognition
requests